sport_league/requirements.txt
django ==3.2.16,<3.3
django-import-export ==3.1.0
django-extensions==3.2.1
django-environ==0.10.0
djangorestframework ==3.14.0
black==23.1.0
isort==5.12.0
pre-commit==3.1.1
pytest==7.2.1
